# contracts/openapi/agent-api.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "GoNow Agent API",
    "version": "1.1.0",
    "description": "Phase 10 runtime-composition repair contract. Production activation remains gated."
  },
  "x-contract-name": "agent-api",
  "x-baseline-status": "phase-10-composition-repair-provisional",
  "x-error-codes": [
    "auth.forbidden",
    "auth.invalid_token",
    "cancel.conflict",
    "context.invalid",
    "idempotency.conflict",
    "internal.error",
    "rate.limit",
    "resume.invalid_or_expired",
    "run.start_rejected",
    "schema.unsupported",
    "service.unavailable",
    "sse.last_event_id_invalid",
    "tenant.scope_missing"
  ],
  "paths": {
    "/health/live": {
      "get": {
        "operationId": "getLiveness",
        "responses": {
          "200": {
            "description": "The process event loop is live.",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HealthProbe"},
                "example": {"status": "live", "reason_codes": []}
              }
            }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "getReadiness",
        "responses": {
          "200": {
            "description": "All mandatory local dependencies are ready.",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HealthProbe"},
                "example": {"status": "ready", "reason_codes": []}
              }
            }
          },
          "503": {
            "description": "At least one mandatory dependency is unavailable.",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HealthProbe"},
                "example": {"status": "not_ready", "reason_codes": ["dependency.database_unavailable"]}
              }
            }
          }
        }
      }
    },
    "/v1/contracts/{name}/{major}": {
      "get": {
        "operationId": "getContractDescriptor",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "name", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[a-z][a-z0-9-]+$"}},
          {"name": "major", "in": "path", "required": true, "schema": {"type": "integer", "minimum": 1}}
        ],
        "responses": {
          "200": {
            "description": "A supported immutable contract descriptor.",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ContractDescriptor"},
                "example": {"name": "agent-api", "major": 1, "version": "1.0.0", "spec_sha256": "0000000000000000000000000000000000000000000000000000000000000000"}
              }
            }
          },
          "401": {"$ref": "#/components/responses/AuthenticationError"},
          "404": {"$ref": "#/components/responses/UnsupportedSchema"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/v1/runs": {
      "post": {
        "operationId": "startRun",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"type": "string", "minLength": 16, "maxLength": 128}}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunStartRequest"}}}
        },
        "responses": {
          "202": {"description": "A durable Run and Job were created or replayed atomically.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunStartResponse"}}}},
          "401": {"$ref": "#/components/responses/AuthenticationError"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/v1/runs/{run_id}/candidate": {
      "get": {
        "operationId": "getRunCandidate",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "run_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "responses": {
          "200": {"description": "The authenticated owner's durable typed Candidate.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ItineraryCandidate"}}}},
          "401": {"$ref": "#/components/responses/AuthenticationError"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/v1/runs/{run_id}/events": {
      "get": {
        "operationId": "streamRunEvents",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "run_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
          {"name": "Last-Event-ID", "in": "header", "required": false, "schema": {"type": "string", "pattern": "^[1-9][0-9]*$"}}
        ],
        "responses": {
          "200": {
            "description": "Owner-scoped at-least-once SSE replay followed by live Events.",
            "headers": {
              "Cache-Control": {"schema": {"type": "string", "const": "no-cache, no-transform"}}
            },
            "content": {"text/event-stream": {"schema": {"type": "string"}}}
          },
          "400": {"$ref": "#/components/responses/InvalidContext"},
          "401": {"$ref": "#/components/responses/AuthenticationError"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/v1/runs/{run_id}/resume": {
      "post": {
        "operationId": "resumeRun",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "run_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResumeRequest"}}}
        },
        "responses": {
          "200": {"description": "The exact one-use capability was consumed.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResumeResponse"}}}},
          "401": {"$ref": "#/components/responses/AuthenticationError"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/Conflict"}
        }
      }
    },
    "/v1/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancelRun",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "run_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CancelRequest"}}}
        },
        "responses": {
          "200": {"description": "Cancellation intent is durable or already terminal.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CancelResponse"}}}},
          "401": {"$ref": "#/components/responses/AuthenticationError"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/Conflict"}
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "schemas": {
      "HealthProbe": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status", "reason_codes"],
        "properties": {
          "status": {"type": "string", "enum": ["live", "ready", "not_ready"]},
          "reason_codes": {"type": "array", "items": {"type": "string"}, "maxItems": 8}
        }
      },
      "ContractDescriptor": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "major", "version", "spec_sha256"],
        "properties": {
          "name": {"type": "string", "const": "agent-api"},
          "major": {"type": "integer", "const": 1},
          "version": {"type": "string", "const": "1.1.0"},
          "spec_sha256": {"type": "string", "pattern": "^[a-f0-9]{64}$"}
        }
      },
      "ItineraryPlanningInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["schema_version", "origin", "destination", "starts_on", "days", "budget_minor", "currency", "locale", "timezone", "hard_constraints"],
        "properties": {
          "schema_version": {"const": "1.0"},
          "origin": {"type": "string", "minLength": 1, "maxLength": 160},
          "destination": {"type": "string", "minLength": 1, "maxLength": 160},
          "starts_on": {"type": "string", "format": "date"},
          "days": {"type": "integer", "minimum": 1, "maximum": 31},
          "budget_minor": {"type": "integer", "minimum": 0, "maximum": 1000000000},
          "currency": {"type": "string", "pattern": "^[A-Z]{3}$"},
          "locale": {"type": "string", "pattern": "^[a-z]{2}(?:-[A-Z]{2})?$"},
          "timezone": {"type": "string", "minLength": 1, "maxLength": 64, "pattern": "^[A-Za-z0-9_+./-]+$"},
          "hard_constraints": {"type": "array", "maxItems": 20, "items": {"type": "string", "minLength": 1, "maxLength": 160}}
        }
      },
      "RunStartRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["schema_version", "thread_id", "itinerary"],
        "properties": {
          "schema_version": {"const": "1.0"},
          "thread_id": {"type": "string", "format": "uuid"},
          "itinerary": {"$ref": "#/components/schemas/ItineraryPlanningInput"}
        }
      },
      "RunStartResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["run_id", "thread_id", "state", "version", "replayed", "behavior_digest"],
        "properties": {
          "run_id": {"type": "string", "format": "uuid"},
          "thread_id": {"type": "string", "format": "uuid"},
          "state": {"const": "queued"},
          "version": {"type": "integer", "minimum": 1},
          "replayed": {"type": "boolean"},
          "behavior_digest": {"type": "string", "pattern": "^[0-9a-f]{64}$"}
        }
      },
      "CandidateCitation": {
        "type": "object",
        "additionalProperties": false,
        "required": ["claim_id", "evidence_id", "source_ref", "sha256"],
        "properties": {
          "claim_id": {"type": "string", "minLength": 1, "maxLength": 128},
          "evidence_id": {"type": "string", "minLength": 1, "maxLength": 128},
          "source_ref": {"type": "string", "pattern": "^evidence://[^\\s]{1,500}$"},
          "sha256": {"type": "string", "pattern": "^[0-9a-f]{64}$"}
        }
      },
      "CandidateItem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["item_id", "title", "start_minute", "duration_minutes", "claim_ids"],
        "properties": {
          "item_id": {"type": "string", "pattern": "^item_[a-z0-9_-]{1,80}$"},
          "title": {"type": "string", "minLength": 1, "maxLength": 160},
          "start_minute": {"type": "integer", "minimum": 0, "maximum": 1439},
          "duration_minutes": {"type": "integer", "minimum": 1, "maximum": 1440},
          "claim_ids": {"type": "array", "items": {"type": "string", "minLength": 1, "maxLength": 128}}
        }
      },
      "CandidateDay": {
        "type": "object",
        "additionalProperties": false,
        "required": ["day_number", "items"],
        "properties": {
          "day_number": {"type": "integer", "minimum": 1, "maximum": 31},
          "items": {"type": "array", "minItems": 1, "maxItems": 30, "items": {"$ref": "#/components/schemas/CandidateItem"}}
        }
      },
      "ItineraryCandidate": {
        "type": "object",
        "additionalProperties": false,
        "required": ["schema_version", "candidate_id", "run_id", "behavior_digest", "input_digest", "title", "days", "citations", "evidence_refs", "status"],
        "properties": {
          "schema_version": {"const": "1.0"},
          "candidate_id": {"type": "string", "pattern": "^cand_[0-9a-f]{32}$"},
          "run_id": {"type": "string", "format": "uuid"},
          "behavior_digest": {"type": "string", "pattern": "^[0-9a-f]{64}$"},
          "input_digest": {"type": "string", "pattern": "^[0-9a-f]{64}$"},
          "title": {"type": "string", "minLength": 1, "maxLength": 160},
          "days": {"type": "array", "minItems": 1, "maxItems": 31, "items": {"$ref": "#/components/schemas/CandidateDay"}},
          "citations": {"type": "array", "items": {"$ref": "#/components/schemas/CandidateCitation"}},
          "evidence_refs": {"type": "array", "uniqueItems": true, "items": {"type": "string", "pattern": "^evidence://[^\\s]{1,500}$"}},
          "status": {"const": "candidate"}
        }
      },
      "ResumeRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["resume_token", "interrupt_id", "command_hash", "command_version"],
        "properties": {
          "resume_token": {"type": "string", "minLength": 1, "writeOnly": true},
          "interrupt_id": {"type": "string", "format": "uuid"},
          "command_hash": {"type": "string", "pattern": "^[0-9a-f]{64}$"},
          "command_version": {"type": "string", "pattern": "^[0-9]+\\.[0-9]+$"}
        }
      },
      "ResumeResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status", "run_id", "interrupt_id", "command_version"],
        "properties": {
          "status": {"const": "resumed"},
          "run_id": {"type": "string", "format": "uuid"},
          "interrupt_id": {"type": "string", "format": "uuid"},
          "command_version": {"type": "string"}
        }
      },
      "CancelRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["expected_version"],
        "properties": {"expected_version": {"type": "integer", "minimum": 0}}
      },
      "CancelResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status", "run_id", "version", "replayed"],
        "properties": {
          "status": {"type": "string", "enum": ["cancelling", "cancelled"]},
          "run_id": {"type": "string", "format": "uuid"},
          "version": {"type": "integer", "minimum": 0},
          "replayed": {"type": "boolean"}
        }
      },
      "PublicError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message", "request_id"],
        "properties": {
          "code": {"type": "string", "enum": ["auth.forbidden", "auth.invalid_token", "cancel.conflict", "context.invalid", "idempotency.conflict", "internal.error", "rate.limit", "resume.invalid_or_expired", "run.start_rejected", "schema.unsupported", "service.unavailable", "sse.last_event_id_invalid", "tenant.scope_missing"]},
          "message": {"type": "string", "minLength": 1, "maxLength": 160},
          "request_id": {"type": "string", "minLength": 1, "maxLength": 128},
          "retry_after_seconds": {"type": "integer", "minimum": 1, "maximum": 86400}
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {"error": {"$ref": "#/components/schemas/PublicError"}}
      }
    },
    "responses": {
      "AuthenticationError": {
        "description": "Authentication failed.",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}, "example": {"error": {"code": "auth.invalid_token", "message": "Authentication failed.", "request_id": "req-example"}}}}
      },
      "Forbidden": {
        "description": "The authenticated principal cannot access the Run.",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}, "example": {"error": {"code": "auth.forbidden", "message": "This action is not allowed.", "request_id": "req-example"}}}}
      },
      "InvalidContext": {
        "description": "The cursor or request context is invalid.",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}, "example": {"error": {"code": "context.invalid", "message": "The request context is invalid.", "request_id": "req-example"}}}}
      },
      "Conflict": {
        "description": "The capability or Run version no longer matches.",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}, "example": {"error": {"code": "context.invalid", "message": "The request context is invalid.", "request_id": "req-example"}}}}
      },
      "UnsupportedSchema": {
        "description": "The requested schema major is unsupported.",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}, "example": {"error": {"code": "schema.unsupported", "message": "The requested schema is unsupported.", "request_id": "req-example"}}}}
      },
      "ServiceUnavailable": {
        "description": "The registry is unavailable.",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}, "example": {"error": {"code": "service.unavailable", "message": "The service is temporarily unavailable.", "request_id": "req-example"}}}}
      }
    }
  }
}